Enhanced exception handling for resultsAPI

+++ src/main/java/com/example/demo/result/ResultController.java
@@ -1,8 +1,10 @@
 package com.example.demo.result;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.orm.jpa.JpaObjectRetrievalFailureException;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -14,8 +16,21 @@ public class ResultController {
     @Autowired
     private ResultService resultService;
     @RequestMapping("/results")
-    public List<Result> getAllResults(){
-        return resultService.getAllResults();
+    public ResponseEntity<List<Result>> getAllResults(){
+        try {
+            List<Result> result = resultService.getAllResults();
+            if (result.size() <= 0)
+            {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User Result not found");
+            }
+            return ResponseEntity.ok(result);
+        }
+        catch (Exception e) {
+            // Handle the exception and return an appropriate response
+            // You can customize the error message based on the exception type
+            return ResponseEntity.status(HttpStatus.NO_CONTENT)
+                    .body(new ArrayList<>());
+        }
     }
 
     @RequestMapping("/results/user/{userId}")
@@ -76,6 +91,14 @@ public class ResultController {
             resultService.addResult(user);
             return ResponseEntity.status(HttpStatus.CREATED).body("Result Added");
         }
+        catch (DataIntegrityViolationException e)
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Result already exists or User/Course doesn't exist");
+        }
+        catch (JpaObjectRetrievalFailureException e)
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("User or course doesn't exist");
+        }
         catch (Exception e)
         {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Error in adding a new result");
@@ -83,11 +106,19 @@ public class ResultController {
 
     }
 
-    @RequestMapping(method = RequestMethod.PUT, value = "/results/{id}")
-    public ResponseEntity<String> updateResult(@RequestBody Result user, @PathVariable int id){
+    @RequestMapping(method = RequestMethod.PUT, value = "/results/")
+    public ResponseEntity<String> updateResult(@RequestBody Result user){
         try {
-            resultService.updateResult(user, id);
+            resultService.updateResult(user);
             return ResponseEntity.status(HttpStatus.OK).body("Result Saved");
+        }
+        catch (DataIntegrityViolationException e)
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Result already exists for the course and the user");
+        }
+        catch (JpaObjectRetrievalFailureException e)
+        {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("User or course doesn't exist");
         }
         catch (Exception e)
         {
@@ -98,8 +129,17 @@ public class ResultController {
     @RequestMapping(method = RequestMethod.DELETE, value = "/results/{id}")
     public ResponseEntity<String> deleteResult(@PathVariable int id){
         try{
-            resultService.deleteResult(id);
-            return ResponseEntity.status(HttpStatus.OK).body("Result Deleted");
+            //Check for getResultId
+            ResponseEntity<Result> response= getResult(id);
+            HttpStatus status = (HttpStatus) response.getStatusCode();
+            if(status == HttpStatus.OK){
+                resultService.deleteResult(id);
+                return ResponseEntity.status(HttpStatus.OK).body("Result Deleted");
+            }
+            else{
+                return ResponseEntity.status(HttpStatus.NO_CONTENT).body("No result found");
+            }
+
         }
         catch (Exception e)
         {
